pyproject.toml
[project]
name = "tradingstrategies"
version = "0.1.0"
description = ""
authors = [
    {name = "Mayuresh Anand",email = "56501614+mayureshucsb2019@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
